requirements.txt

asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.1
joblib==1.1.0
Markdown==3.3.6
networkx==2.6.3
Pillow==9.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
zipp==3.7.0
gunicorn==20.1.0
whitenoise==5.3.0

